[jenkins] - update deploy stage

+++ Jenkinsfile
@@ -12,9 +12,14 @@ pipeline {
         stage('Deploy Docker Container') {
             steps {
                 script {
-                    sh "docker stop blog"
-                    sh "docker rm blog"
-                    sh "docker run -d --name blog --mount source=/usr/share/nginx/html/quasar, target=/usr/share/nginx/html/quasar --network=nginx_infra blog"
+                    def containerExists = sh(script: 'docker ps -aqf name=blog', returnStdout: true).trim()
+
+                    if (containerExists) {
+                        sh "docker stop blog"
+                        sh "docker rm blog"
+                    }
+
+                    sh "docker run -d --name blog --network=nginx_infra blog"
                 }
             }
         }
